+++ src/canns/models/basic/__init__.py
@@ -1,4 +1,11 @@
-from .cann import CANN1D, CANN1D_SFA, CANN2D, CANN2D_SFA
+from .cann import (
+    ACCL_DEFAULT_K,
+    ACCL_MODES,
+    CANN1D,
+    CANN1D_SFA,
+    CANN2D,
+    CANN2D_SFA,
+)
 from .grid_cell import GridCell2DPosition, GridCell2DVelocity
 from .hierarchical_model import HierarchicalNetwork
 
@@ -10,4 +17,7 @@ __all__ = [
     "GridCell2DPosition",
     "GridCell2DVelocity",
     "HierarchicalNetwork",
+    # Low-rank acceleration of the recurrent matvec
+    "ACCL_MODES",
+    "ACCL_DEFAULT_K",
 ]